# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tenet Audit — EU AI Act Audit Log API",
    "description": "Tamper-evident AI call logging, risk classification (Article 6 / Annex III), incident reporting (Article 73) and auditor-ready evidence packs (Article 12).",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Tenant key (ta_live_…) or admin key (ta_admin_…)"
      }
    },
...
  },
  "paths": {
...
    "/v1/admin/tenants": {
      "post": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "slug"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "slug": {
                    "pattern": "^[a-z0-9][a-z0-9-]{1,62}[a-z0-9]$",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "tenant",
                    "apiKey",
                    "genesisHash"
                  ],
                  "properties": {
                    "tenant": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "slug",
                        "status",
                        "createdAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      }
                    },
                    "apiKey": {
                      "type": "object",
                      "required": [
                        "id",
                        "secret",
                        "keyPrefix"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "secret": {
                          "description": "Shown exactly once — store it securely.",
                          "type": "string"
                        },
                        "keyPrefix": {
                          "type": "string"
                        }
                      }
                    },
                    "genesisHash": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "slug",
                      "status",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/tenants/{tenantId}/api-keys": {
      "post": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "maxLength": 200,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "tenantId",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "secret",
                    "keyPrefix"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "secret": {
                      "type": "string"
                    },
                    "keyPrefix": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/api-keys/{keyId}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "keyId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ],
...
}